Add new file

+++ c_to_f.py
@@ -1,4 +1,4 @@
 c = input('请输入摄氏温度： ')
-c = int(c)
+c = float(c)
 f = c * 9 / 5 + 32
 print('华氏温度是： ', f)
